Conflicts:
	.travis.yml

+++ src/ununpack/agent_tests/Functional/cliParamsTest4UnunpackException.php
@@ -1,7 +1,7 @@
 <?php
-
 /*
  Copyright (C) 2010-2012 Hewlett-Packard Development Company, L.P.
+ Copyright (C) 2015 Siemens AG
 
  This program is free software; you can redistribute it and/or
  modify it under the terms of the GNU General Public License
@@ -25,43 +25,69 @@
  */
 require_once './utility.php';
 
-class cliParamsTest4UnunpackExcption extends PHPUnit_Framework_TestCase {
+use Fossology\Lib\Test\TestPgDb;
+use Fossology\Lib\Test\TestInstaller;
+
+class cliParamsTest4UnunpackExcption extends PHPUnit_Framework_TestCase
+{
+  private $agentDir;
+
+  /** @var TestPgDb */
+  private $testDb;
+  /** @var TestInstaller */
+  private $testInstaller;
+
+  function setUp()
+  {
+    $this->testDb = new TestPgDb('ununpackExceptional');
+    $this->agentDir = dirname(dirname(__DIR__))."/";
+
+    $sysConf = $this->testDb->getFossSysConf();
+    $this->testInstaller = new TestInstaller($sysConf);
+    $this->testInstaller->init();
+    $this->testInstaller->install($this->agentDir);
+
+    $this->testDb->createSequences(array(), true);
+    $this->testDb->createPlainTables(array(), true);
+    $this->testDb->alterTables(array(), true);
+  }
+
+  public function tearDown()
+  {
+    $this->testInstaller->uninstall($this->agentDir);
+    $this->testInstaller->clear();
+    $this->testInstaller->rmRepo();
+    $this->testDb = null;
+  }
+
   /* command is */
   public function testValidParam(){
-    print "test unpack with an invalid parameter\n";
     global $UNUNPACK_CMD;
     global $TEST_DATA_PATH;
     global $TEST_RESULT_PATH;
-    global $fossology_testconfig;
-    
-    //$UNUNPACK_CMD = "../../agent/ununpack";
-    $fossology_testconfig = getenv('FOSSOLOGY_TESTCONFIG');
-    $UNUNPACK_CMD = $fossology_testconfig . "/mods-enabled/ununpack/agent/ununpack";
-    exec("/bin/rm -rf $TEST_RESULT_PATH");
+    $fossology_testconfig = $this->testDb->getFossSysConf();
+
+    $UNUNPACK_CMD = $this->agentDir . "/agent/ununpack";
+    if (!empty($TEST_RESULT_PATH))
+      exec("/bin/rm -rf $TEST_RESULT_PATH");
     $command = "$UNUNPACK_CMD -qCRs $TEST_DATA_PATH/523.iso -d $TEST_RESULT_PATH -c $fossology_testconfig > /dev/null 2>&1";
     $last = exec($command, $usageOut, $rtn);
     $this->assertNotEquals($rtn, 0);
     $this->assertFileNotExists("$TEST_RESULT_PATH/523.iso.dir/523SFP/QMFGOEM.TXT");
   }
-   
+
   /* test null-file */
   public function testNullFile(){
-    print "test unpack with an null file\n";
-    global $UNUNPACK_CMD;
     global $TEST_DATA_PATH;
     global $TEST_RESULT_PATH;
-    global $fossology_testconfig;
+    $fossology_testconfig = $this->testDb->getFossSysConf();
 
-    //$UNUNPACK_CMD = "../../agent/ununpack";
-    $fossology_testconfig = getenv('FOSSOLOGY_TESTCONFIG');
-    $UNUNPACK_CMD = $fossology_testconfig . "/mods-enabled/ununpack/agent/ununpack";
-    exec("/bin/rm -rf $TEST_RESULT_PATH");
+    $UNUNPACK_CMD = $this->agentDir . "/agent/ununpack";
+    if (!empty($TEST_RESULT_PATH))
+      exec("/bin/rm -rf $TEST_RESULT_PATH");
     $command = "$UNUNPACK_CMD -qCR $TEST_DATA_PATH/null-file -d $TEST_RESULT_PATH -c $fossology_testconfig > /dev/null 2>&1";
     $last = exec($command, $usageOut, $rtn);
     $this->assertNotEquals($rtn, 0);
     $this->assertFileNotExists("$TEST_RESULT_PATH/null-file.dir/");
   }
 }
-
-
-?>
